Wrote GameStartDataSender and handlers for game creation messages.

+++ castledice-game-server/NetworkManager/PlayersDictionary.cs
@@ -5,5 +5,5 @@
 /// </summary>
 public static class PlayersDictionary
 {
-    public static Dictionary<int, int> Dictionary { get; } = new();
+    public static Dictionary<int, ushort> Dictionary { get; } = new();
 }
